requirements.txt
openai>=1.0.0
pillow>=10.0.0
opencv-python>=4.8.0
requests>=2.31.0
jinja2>=3.1.0
python-dotenv>=1.0.0
flask>=3.0.0
numpy>=1.24.0
